# FreeBSD: convert teardown inactive lock to a read-mostly sleepable lock