@freqlabs 2020-09-09
@mjguzik no, you're good.